backend/requirements.txt
fastapi
uvicorn
openai
langchain
chromadb
python-dotenv
requests
tinydb

